changing some ints to size_t and adjusting a strncpy to have room for a null term

+++ earth_enterprise/src/fusion/gst/gstExport.cpp
@@ -499,7 +499,7 @@ bool gstPacketFileExporter::AddLandmark(etLandmarkPacketData* pakdata,
   assert(desc != NULL);
   if (!desc->IsEmpty()) {
     std::string src = desc->ValueAsString();
-    int sz = src.length();
+    size_t sz = src.length();
     char * ptr = static_cast<char*>(mem_pool_.Allocate(sz + 1));
     strncpy(ptr, src.c_str(), sz + 1);
     while (sz && isspace(ptr[sz - 1])) {
